minor refactoring + relative path test
