+++ packages/notest-monitor/src/function-wrapper/info-adder-for-collector.ts
