+++ packages/notest-monitor/src/function-wrapper/statements_instrumenters/return_instrumenter.ts
@@ -1,6 +1,6 @@
 import {FunctionDeclaration, ReturnStatement, Statement, SyntaxKind} from "ts-morph"
 import {InstrumentStatementInterface} from "./instrument_statement.interface";
-import {InfoAdderForCollector} from "../info_adder_for_collector";
+import {InfoAdderForCollector} from "../info-adder-for-collector";
 
 export class ReturnInstrumenter implements InstrumentStatementInterface {
